bug fix create_insee_folder

+++ R/clean_insee_folder.R
@@ -1,7 +1,7 @@
 #' @noRd
 clean_insee_folder = function(){
 
-  dir_creation_fail = try(create_insee_dir(), silent = TRUE)
+  dir_creation_fail = try(create_insee_folder(), silent = TRUE)
 
   if(!"try-error" %in% class(dir_creation_fail)){
     insee_local_dir = rappdirs::user_data_dir("insee")
@@ -9,13 +9,18 @@ clean_insee_folder = function(){
     insee_local_dir = tempdir()
   }
 
-  list_file_insee = file.path(insee_local_dir,
-                              list.files(insee_local_dir))
+  if(insee_local_dir == rappdirs::user_data_dir("insee")){
 
-  if(length(list_file_insee) > 0){
-    for(file_name in list_file_insee){
-      try(file.remove(file_name), silent = TRUE)
+    list_file_insee = file.path(insee_local_dir,
+                                list.files(insee_local_dir))
+
+    if(length(list_file_insee) > 0){
+      for(file_name in list_file_insee){
+        try(file.remove(file_name), silent = TRUE)
+      }
     }
+
   }
+
 }
